STS credentials are fetched from STS regional endpoint instead of global endpoint
As per AWS STS recommendation, STS credentials for provided roleARN are
fetched from STS regional endpoint instead of global endpoint. If the
configured region is disabled for the STS token, the credentials are
fetched from primary regional endpoint in that AWS partition. The
primary regional endpoint cannot be disabled for STS token.